handle statue code by exception | 通过异常来处理状态码

+++ src/main/java/com/example/demo/user/api/UserApiController.java
@@ -1,8 +1,11 @@
 package com.example.demo.user.api;
 
+import com.example.demo.error.EntityNotFoundException;
 import com.example.demo.user.User;
 import com.example.demo.user.UserRepository;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
@@ -13,27 +16,40 @@ public class UserApiController {
     private UserRepository userRepository;
 
     @Autowired
-    public UserApiController(UserRepository userRepository){
-        this.userRepository=userRepository;
+    public UserApiController(UserRepository userRepository) {
+        this.userRepository = userRepository;
     }
 
-    @RequestMapping(value = "/users",method = RequestMethod.GET)
-    public List<User> findAll(){
+    @RequestMapping(value = "/users", method = RequestMethod.GET)
+    public List<User> findAll() {
         return userRepository.findAll();
     }
 
-    @RequestMapping(value = "/users",method = RequestMethod.POST)
-    public User createUser(@RequestBody User user){
-        return userRepository.save(user);
+    @RequestMapping(value = "/users", method = RequestMethod.POST)
+    public ResponseEntity<User> createUser(@RequestBody User user) {
+        HttpStatus status = HttpStatus.OK;
+        if (!userRepository.exists(user.getEmail())) {
+            status = HttpStatus.CREATED;
+        }
+        user = userRepository.save(user);
+        return new ResponseEntity<>(user, status);
     }
 
-    @RequestMapping(value = "/user/{email}",method = RequestMethod.PUT)
-    public User updateUser(@PathVariable String email,@RequestBody User user){
-        return userRepository.save(email,user);
+    @RequestMapping(value = "/user/{email}", method = RequestMethod.PUT)
+    public ResponseEntity<User> updateUser(@PathVariable String email, @RequestBody User user) throws EntityNotFoundException {
+        if (!userRepository.exists(user.getEmail())) {
+            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+        }
+        user = userRepository.update(email, user);
+        return new ResponseEntity<>(user,HttpStatus.CREATED);
     }
 
-    @RequestMapping(value = "/user/{email}",method = RequestMethod.DELETE)
-    public void deleteUser(@PathVariable String email){
+    @RequestMapping(value = "/user/{email}", method = RequestMethod.DELETE)
+    public ResponseEntity<User> deleteUser(@PathVariable String email) throws EntityNotFoundException {
+        if(!userRepository.exists(email)){
+            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+        }
         userRepository.delete(email);
+        return new ResponseEntity<>(HttpStatus.OK);
     }
 }
